moved merge_eopatch to a separate module and minor improvements
