+++ core/eolearn/core/constants.py
@@ -186,7 +186,6 @@ class OverwritePermission(Enum):
     - `OVERWRITE_FEATURES` - Overwrite only data for features which have to be saved. The remaining content of saved
         EOPatch will stay unchanged.
     - `OVERWRITE_PATCH` - Overwrite entire content of saved EOPatch and replace it with the new content.
-    - `MERGE_FEATURES` - Combine new and old content of EOPatch.
     """
     ADD_ONLY = 0
     OVERWRITE_FEATURES = 1
